tweaks to layout stories
