+++ src/components/layout/App.svelte
@@ -2,18 +2,20 @@
 import { setContext } from 'svelte';
 import { writable } from 'svelte/store';
 
-export let layoutType = 'vertical';
+export let orientation = 'vertical';
 
 const layout = writable({});
-$layout.orientation = layoutType;
+
+$layout.orientation = orientation;
+
 setContext('_layout', layout);
 
 </script>
 
 <div 
   class=app-layout
-  class:app-layout--vertical={layoutType === 'vertical'}
-  class:app-layout--horizontal={layoutType === 'horizontal'}
+  class:app-layout--vertical={orientation === 'vertical'}
+  class:app-layout--horizontal={orientation === 'horizontal'}
 >
   <slot></slot>
 </div>
